refactor(iperf): decompose the installer god-file by role (WS-D)
Split internal/diagnostics/iperf/installer.go (698 LOC) into role-based
files in the same package — pure move: package-manager detection and
build-dependency reporting stay in installer.go; install/build execution
(package-manager, GitHub source, download) moves out.

All files under the 600-line size-gate threshold. Declaration sets and
code-line multisets verified identical to the original.

Related to #187
